Apply design pattern to booking and payment component Implement add card to payment of user

+++ src/main/java/mum/asd/repository/CardRepository.java
@@ -1,9 +1,10 @@
 package mum.asd.repository;
 
-import mum.asd.domain.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import mum.asd.domain.Card;
+
 @Repository
-public interface CardRepository extends JpaRepository<Card, Integer> {
+public interface CardRepository extends JpaRepository<Card, Long> {
 }
